RavenDB-18987 - Optimize entries deletes in Corax - For Small sets, we'll update the value in place (instead of rebuilding from scratch) - Modify everything that work with small sets to account for the size != the number of items - Entries count should be long, we now support > 2b records

+++ src/Raven.Server/Documents/Indexes/Index.cs
@@ -2273,7 +2273,7 @@ namespace Raven.Server.Documents.Indexes
 
                     try
                     {
-                        int? entriesCount = null;
+                        long? entriesCount = null;
 
                         using (InitializeIndexingWork(indexContext))
                         {
